Rename parse script functions
